# ARROW-10010: [Rust] Speedup arithmetic (1.3-1.9x)